Deprecate ElseOnExhaustiveCase diagnostic
