+++ sig/steep/diagnostic/ruby.rbs
@@ -475,14 +475,6 @@ module Steep
         def header_line: () -> ::String
       end
 
-      class ElseOnExhaustiveCase < Base
-        attr_reader type: untyped
-
-        def initialize: (node: untyped, type: untyped) -> void
-
-        def header_line: () -> String
-      end
-
       class UnreachableBranch < Base
         def header_line: () -> String
       end
